desktop: Move locale-related stuff to gui/locale.rs
This refactors moves locale-related code away from
generic GUI-related code into a separate file.
